+++ src/components/homepage/FilterSidebar.js
@@ -23,7 +23,7 @@ export default function FilterSidebar() {
     const classes = useStyles()
 
     return (
-        <div   style={{display: 'block', /*position: '-webkit-sticky',*/ position: 'sticky', top: "104px", height: '85vh', border: '1px solid rgb(128, 128, 128, 0.12)', overflowY: 'auto', borderRadius: "4px" }}>
+        <div   style={{display: 'block', position: '-webkit-sticky', top: "104px", height: '85vh', border: '1px solid rgb(128, 128, 128, 0.12)', overflowY: 'auto', borderRadius: "4px" }}>
             <Paper elevation={2} className={classes.root}>
             <Typography variant="h3" className={classes.title}>
                 Filter your search
