+++ src/pages/CreatePost/CreatePostForm.js
@@ -35,6 +35,7 @@ const useStyles = makeStyles((theme) => ({
   },
   stepper: {
     padding: theme.spacing(3, 0, 5),
+    background: "inherit"
   },
 }))
